refactor: extracts bugsnag configuration
