+++ src/index.tsx
@@ -13,39 +13,13 @@ import SearchHistoryContextProvider from './search-history';
 import TicketContextProvider from './TicketContext';
 import RemoteConfigContextProvider from './RemoteConfigContext';
 import {loadLocalConfig} from './local-config';
-import Bugsnag, {Event} from '@bugsnag/react-native';
+import Bugsnag from '@bugsnag/react-native';
 import {setInstallId as setApiInstallId} from './api/client';
 import ErrorBoundary from './error-boundary';
 import {PreferencesContextProvider} from './preferences';
+import configureAndStartBugsnag from './diagnostics/bugsnagConfig';
 
-if (!__DEV__) {
-  Bugsnag.start();
-} else {
-  Bugsnag.notify = function (error, onError) {
-    const event = Event.create(error, true, {} as any, 'notify()', 1, console);
-    let metadata: {[key: string]: any} = {};
-    event.addMetadata = (
-      section: string,
-      values: string | object,
-      values2?: unknown,
-    ) => {
-      if (typeof values === 'string') {
-        metadata[values] = values2;
-      } else {
-        metadata = {
-          ...metadata,
-          ...values,
-        };
-      }
-    };
-    onError?.(event);
-    console.error('[BUGSNAG]', error, JSON.stringify(metadata, null, 2));
-  };
-  Bugsnag.leaveBreadcrumb = (message, metadata) =>
-    // eslint-disable-next-line
-    console.log(message, metadata);
-  Bugsnag.setUser = () => {};
-}
+configureAndStartBugsnag();
 
 import {MAPBOX_API_TOKEN} from '@env';
 import MapboxGL from '@react-native-mapbox-gl/maps';
